+++ malachite-nz/src/bin_util/demo_and_bench/gaussian_integer/factorization/mod.rs
@@ -6,5 +6,10 @@
 // Lesser General Public License (LGPL) as published by the Free Software Foundation; either version
 // 3 of the License, or (at your option) any later version. See <https://www.gnu.org/licenses/>.
 
-pub mod arithmetic;
-pub mod factorization;
+use malachite_base::test_util::runner::Runner;
+
+pub(crate) fn register(runner: &mut Runner) {
+    remove_one_plus_i::register(runner);
+}
+
+mod remove_one_plus_i;