+++ malachite-nz/src/gaussian_integer/factorization/mod.rs
@@ -6,5 +6,6 @@
 // Lesser General Public License (LGPL) as published by the Free Software Foundation; either version
 // 3 of the License, or (at your option) any later version. See <https://www.gnu.org/licenses/>.
 
-pub mod arithmetic;
-pub mod factorization;
+/// Functions for removing the largest power of $1 + i$, the prime above 2, from a
+/// [`GaussianInteger`](crate::gaussian_integer::GaussianInteger).
+pub mod remove_one_plus_i;