+++ malachite-nz/src/test_util/gaussian_integer/factorization/mod.rs
@@ -6,5 +6,4 @@
 // Lesser General Public License (LGPL) as published by the Free Software Foundation; either version
 // 3 of the License, or (at your option) any later version. See <https://www.gnu.org/licenses/>.
 
-pub mod arithmetic;
-pub mod factorization;
+pub mod remove_one_plus_i;